+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 
 CHANGES:
 
+1. Add write capability to socket_stream wrapper, allowing clients to write to NMEAReader socket stream (NMEAReader.datastream) as well as read from it.
 1. Update constructor arguments and docstrings to clarify API (no functional changes).
 
 ### RELEASE 1.0.26